@@ -1,6 +1,6 @@
 {
   "name": "@kintone/create-plugin",
-  "version": "1.3.5",
+  "version": "2.0.3",
   "description": "A CLI tool for creating a kintone plugin!",
   "bin": {
     "create-plugin": "bin/cli.js",
@@ -8,38 +8,37 @@
   },
   "main": "dist/src/index.js",
   "engines": {
-    "node": ">=6"
+    "node": ">=8"
   },
   "repository": "kintone/create-plugin",
   "homepage": "https://github.com/kintone/create-plugin",
   "bugs": "https://github.com/kintone/create-plugin/issues",
   "dependencies": {
-    "chalk": "^2.4.1",
-    "glob": "^7.1.3",
-    "inquirer": "^6.2.0",
-    "lodash": "^4.17.11",
+    "chalk": "^2.4.2",
+    "glob": "^7.1.4",
+    "inquirer": "^7.0.0",
+    "lodash": "^4.17.15",
     "meow": "^5.0.0",
-    "node-rsa": "^1.0.1",
-    "os-locale": "^3.0.1",
-    "rimraf": "^2.6.2"
+    "node-rsa": "^1.0.5",
+    "os-locale": "^4.0.0",
+    "rimraf": "^3.0.0"
   },
   "devDependencies": {
-    "@kintone/plugin-packer": "^1.0.4",
     "@types/glob": "^7.1.1",
-    "@types/inquirer": "^0.0.43",
-    "@types/lodash": "^4.14.118",
-    "@types/mocha": "^5.2.5",
-    "@types/node": "^9.6.39",
-    "@types/node-rsa": "^0.4.3",
+    "@types/inquirer": "^6.5.0",
+    "@types/lodash": "^4.14.138",
+    "@types/mocha": "^5.2.7",
+    "@types/node": "^9.6.51",
+    "@types/node-rsa": "^1.0.0",
     "@types/rimraf": "^2.0.2",
-    "mocha": "^5.2.0",
+    "mocha": "^6.2.0",
     "npm-run-all": "^4.1.5",
-    "prettier": "^1.15.2",
-    "standard-version": "^4.4.0",
-    "ts-node": "^7.0.1",
-    "tslint": "^5.11.0",
+    "prettier": "^1.18.2",
+    "standard-version": "^7.0.0",
+    "ts-node": "^8.3.0",
+    "tslint": "^5.19.0",
     "tslint-plugin-prettier": "^2.0.1",
-    "typescript": "^3.1.6"
+    "typescript": "^3.6.2"
   },
   "files": [
     "bin",
@@ -50,8 +49,8 @@
     "start": "tsc -w",
     "build": "tsc",
     "lint": "tslint -c tslint.json src/**/*.ts test/**/*.ts",
-    "release": "run-s build lint test:all standard-version",
-    "standard-version": "standard-version",
+    "release": "run-s build lint test:all && standard-version",
+    "release:dryrun": "standard-version --dry-run",
     "unittest": "mocha --require ts-node/register --watch-extensions ts test/**/*-test.ts",
     "test": "npm-run-all -s build -p lint unittest",
     "test:all": "run-p test e2e",
