@@ -1,24 +1,24 @@
 jQuery.noConflict();
 
 (function($, PLUGIN_ID) {
-    'use strict';
+  'use strict';
 
-    var $form = $('.js-submit-settings');
-    var $message = $('.js-text-message');
+  var $form = $('.js-submit-settings');
+  var $cancelButton = $('.js-cancel-button');
+  var $message = $('.js-text-message');
+  var config = kintone.plugin.app.getConfig(PLUGIN_ID);
 
-    function getSettingsUrl() {
-        return '/k/admin/app/flow?app=' + kintone.app.getId();
-    }
-
-    var config = kintone.plugin.app.getConfig(PLUGIN_ID);
-    if (config.message) {
-        $message.val(config.message);
-    }
-    $form.on('submit', function(e) {
-        e.preventDefault();
-        kintone.plugin.app.setConfig({message: $message.val()}, function() {
-            alert('Please update the app!');
-            window.location.href = getSettingsUrl();
-        });
+  if (config.message) {
+    $message.val(config.message);
+  }
+  $form.on('submit', function(e) {
+    e.preventDefault();
+    kintone.plugin.app.setConfig({message: $message.val()}, function() {
+      alert('The plug-in settings have been saved. Please update the app!');
+      window.location.href = '/k/admin/app/flow?app=' + kintone.app.getId();
     });
+  });
+  $cancelButton.on('click', function() {
+    window.location.href = '/k/admin/app/' + kintone.app.getId() + '/plugin/';
+  });
 })(jQuery, kintone.$PLUGIN_ID);
